feat: add POST transaction endpoint

+++ api/internal/repository/entity/account.go
@@ -1,6 +1,6 @@
 package entity
 
 type Account struct {
-	ID       uint   `gorm:"primaryKey;autoIncrement"`
+	ID             uint64   `gorm:"primaryKey;autoIncrement"`
 	DocumentNumber string `gorm:"column:document_number;unique;not null"`
 }
